Compilation avec Weverything OK ; Update README ; Corrections sur openfile, itoa, itoa_base, lltoa, lstat

+++ ft_memchr.c
@@ -6,11 +6,12 @@
 /*   By: banthony <banthony@students.42.fr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/12/03 11:55:52 by banthony          #+#    #+#             */
-/*   Updated: 2015/12/07 16:58:08 by banthony         ###   ########.fr       */
+/*   Updated: 2017/08/16 22:37:43 by banthony         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <string.h>
+#include "libft.h"
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
@@ -18,7 +19,7 @@ void	*ft_memchr(const void *s, int c, size_t n)
 	size_t			i;
 
 	i = 0;
-	str = (unsigned char *)s;
+	str = (unsigned char *)(unsigned long)s;
 	while (i < n)
 	{
 		if (str[i] != (unsigned char)c)
