+++ ft_memcmp.c
@@ -6,11 +6,12 @@
 /*   By: banthony <banthony@students.42.fr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/12/03 15:35:12 by banthony          #+#    #+#             */
-/*   Updated: 2015/12/07 17:00:18 by banthony         ###   ########.fr       */
+/*   Updated: 2017/08/16 22:37:22 by banthony         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <string.h>
+#include "libft.h"
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
@@ -19,8 +20,8 @@ int	ft_memcmp(const void *s1, const void *s2, size_t n)
 	size_t			i;
 
 	i = 0;
-	str1 = (unsigned char *)s1;
-	str2 = (unsigned char *)s2;
+	str1 = (unsigned char *)(unsigned long)s1;
+	str2 = (unsigned char *)(unsigned long)s2;
 	if (!s1 || !s2 || n == 0)
 		return (0);
 	while (str1[i] == str2[i] && i + 1 < n)
