+++ ft_memcpy.c
@@ -6,11 +6,12 @@
 /*   By: banthony <banthony@students.42.fr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/12/02 19:31:06 by banthony          #+#    #+#             */
-/*   Updated: 2015/12/07 17:01:22 by banthony         ###   ########.fr       */
+/*   Updated: 2017/08/17 00:53:12 by banthony         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <string.h>
+#include "libft.h"
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
@@ -20,7 +21,7 @@ void	*ft_memcpy(void *dest, const void *src, size_t n)
 
 	i = 0;
 	str1 = (unsigned char *)dest;
-	str2 = (unsigned char *)src;
+	str2 = (unsigned char *)(unsigned long)src;
 	while (i < n)
 	{
 		str1[i] = str2[i];
