+++ ft_newtab.c
@@ -6,7 +6,7 @@
 /*   By: banthony <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/01/22 12:24:32 by banthony          #+#    #+#             */
-/*   Updated: 2016/01/22 16:03:10 by banthony         ###   ########.fr       */
+/*   Updated: 2017/08/16 22:29:18 by banthony         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,13 +22,13 @@ char	**ft_newtab(int line, int col, int c)
 	tab = NULL;
 	if (line < 0 || col < 0 || c < 0)
 		return (NULL);
-	if (!(tab = (char **)malloc(sizeof(char *) * line + 1)))
+	if (!(tab = (char **)malloc(sizeof(char *) * (size_t)(line + 1))))
 		return (NULL);
 	while (++i < line)
 	{
-		if (!(tab[i] = (char *)malloc(sizeof(char) * col + 1)))
+		if (!(tab[i] = (char *)malloc(sizeof(char) * (size_t)(col + 1))))
 			return (NULL);
-		ft_memset(tab[i], c, col + 2);
+		ft_memset(tab[i], c, (size_t)(col + 2));
 		tab[i][col] = '\n';
 		tab[i][col + 1] = '\0';
 	}
