+++ ft_strrchr.c
@@ -6,11 +6,12 @@
 /*   By: banthony <banthony@students.42.fr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/12/02 18:48:09 by banthony          #+#    #+#             */
-/*   Updated: 2015/12/02 18:52:17 by banthony         ###   ########.fr       */
+/*   Updated: 2017/08/17 00:52:06 by banthony         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <string.h>
+#include "libft.h"
 
 char	*ft_strrchr(const char *s, int c)
 {
@@ -22,11 +23,11 @@ char	*ft_strrchr(const char *s, int c)
 	while (s[i] != '\0')
 	{
 		if (s[i] == c)
-			ptr = (char *)&s[i];
+			ptr = (char *)(unsigned long)&s[i];
 		i++;
 	}
 	if (s[i] == c)
-		return ((char *)&s[i]);
+		return ((char *)(unsigned long)&s[i]);
 	else
 		return (ptr);
 }
